delete: remove unused test, update styles, and refine documentation
- Deleted `UnitTest1` to clean up unused test files.
- Updated `_FilterBar.cshtml` and `_Layout.cshtml` styles for better consistency and appearance.
- Minor restructuring in `_HomePartial.cshtml` to improve layout usability.
- Added `[TOC]` to `umamiproductionready.md` for better navigation.
- Updated API endpoints in `Umami.Net/README.md` to include HTTP methods for clarity.
- Bumped package versions in `SentimentAnalysis.csproj` to resolve dependency inconsistencies.
- Added `TODO` comments in `Florence2ImageAnalysisService` to reflect breaking API changes requiring fixes.

+++ Mostlylucid.SentimentAnalysis/Services/SentimentAnalysisService.cs
@@ -258,7 +258,7 @@ public class SentimentAnalysisService : ISentimentAnalysisService
 
         // Simple readability metric based on average sentence length and word length
         float avgSentenceLength = (float)words.Count / sentences.Count;
-        float avgWordLength = words.Average(w => w.Length);
+        float avgWordLength = (float)words.Average(w => w.Length);
 
         // Normalize to 0-1 scale (lower values = easier to read)
         float sentenceComplexity = Math.Clamp(1.0f - (avgSentenceLength / 30.0f), 0, 1);
